Flesh out Userdata and use in callback wrappers
